+++ client/src/app/containers/Home/ActiveGamesTable.jsx
@@ -1,20 +1,25 @@
 import React from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { useInjectReducer, useInjectSaga } from 'redux-injectors';
-import { selectCurrentUser, selectActiveGames } from '../../../auth/selectors';
+import { selectCurrentUser } from '../../../auth/selectors';
 import { authSaga } from '../../../auth/saga';
 import { sliceKey, reducer, actions } from '../../../auth/slice';
+import { selectActiveGames } from '../../../home/selectors';
+import { homeSaga } from '../../../home/saga';
+import { sliceKey as homeSliceKey, reducer as homeReducer, actions as homeActions } from '../../../home/slice';
 
 const ActiveGamesTable = () => {
   useInjectReducer({ key: sliceKey, reducer: reducer });
   useInjectSaga({ key: sliceKey, saga: authSaga });
+  useInjectReducer({ key: homeSliceKey, reducer: homeReducer });
+  useInjectSaga({ key: homeSliceKey, saga: homeSaga });
   const dispatch = useDispatch();
   const currentUser = useSelector(selectCurrentUser);
-  const activeGames = useSelector(selectActiveGames);
+  const activeGames = useSelector(selectActiveGames(currentUser.id));
   
   // event handlers
   const onRefreshActiveGames = () => {
-    dispatch(actions.refreshActiveGames({ id: currentUser.id }));
+    dispatch(homeActions.refreshActiveGames({ id: currentUser.id }));
   };
 
   let gameButtons = [];
